split non/kernel chainparamsbase
