+++ src/Makefile.am
@@ -666,6 +666,7 @@ libbitcoin_util_a_SOURCES = \
   interfaces/echo.cpp \
   interfaces/handler.cpp \
   interfaces/init.cpp \
+  kernel/chainparamsbase.cpp \
   logging.cpp \
   random.cpp \
   randomenv.cpp \
@@ -894,6 +895,7 @@ libbitcoinkernel_la_SOURCES = \
   hash.cpp \
   kernel/chain.cpp \
   kernel/checks.cpp \
+  kernel/chainparamsbase.cpp \
   kernel/coinstats.cpp \
   kernel/context.cpp \
   kernel/mempool_persist.cpp \
